Refactor CatPad state into hooks
